refactor: Change pipeline to follow the structure of engine and adapters; Create new types of pipelines.
