+++ filejacket/__init__.py
@@ -21,6 +21,18 @@ Should there be a need for contact the electronic mail
 `filejacket <at> gabrielfontenelle.com` can be used.
 """
 
+from .adapters.image import OpenCVImage, PillowImage, WandImage
+# Module with classes adapted from engines
+from .adapters.mimetype import LibraryMimeTyper, APIMimeTyper
+from .adapters.pipeline import PipelineContent, PipelineOrderedDependency, PipelineSequential, ProcessorContent
+from .adapters.storage import WindowsFileSystem, LinuxFileSystem
+from .adapters.video import MoviePyVideo
+from .engines.image import ImageEngine
+# A Pipeline is a sequence that loop processors to be run.
+from .engines.pipeline import Processor, PipelineEngine
+# Module with engines for adapters
+from .engines.storage import StorageEngine
+from .engines.video import VideoEngine
 from .exception import (
     ImproperlyConfiguredFile,
     NoInternalContentError,
@@ -30,22 +42,6 @@ from .exception import (
 )
 from .file import BaseFile
 from .handler import System, URI
-
-# Module with engines for adapters
-from .engines.storage import StorageEngine
-from .engines.image import ImageEngine
-from .engines.video import VideoEngine
-
-# Module with classes adapted from engines
-from .adapters.mimetype import LibraryMimeTyper, APIMimeTyper
-from .adapters.image import OpenCVImage, PillowImage, WandImage
-from .adapters.video import MoviePyVideo
-from .adapters.storage import WindowsFileSystem, LinuxFileSystem
-
-# Module with classes that define the pipelines and its processors classes.
-# A Pipeline is a sequence that loop processors to be run.
-from .pipelines import Processor, Pipeline
-
 # Module with base classes for pipeline classes.
 from .pipelines.base import (
     BaseComparer,
@@ -54,7 +50,6 @@ from .pipelines.base import (
     BaseRenamer,
     BaseRender,
 )
-
 # Module with pipeline classes for comparing Files.
 from .pipelines.comparer import (
     BinaryCompare,
@@ -66,7 +61,6 @@ from .pipelines.comparer import (
     SizeCompare,
     TypeCompare,
 )
-
 # Module with pipeline classes for extracting data for files from multiple sources.
 from .pipelines.extractor import (
     PackageExtractor,
@@ -84,13 +78,10 @@ from .pipelines.extractor import (
     ZipCompressedFilesFromPackageExtractor,
     MimeTypeFromContentExtractor,
 )
-
 # Module with pipeline classes for generating or extracting hashed data related to file.
 from .pipelines.hasher import CRC32Hasher, MD5Hasher, SHA256Hasher
-
 # Module with pipeline classes for renaming files.
 from .pipelines.renamer import WindowsRenamer, LinuxRenamer, UniqueRenamer
-
 # module with pipeline classes for render content representation.
 from .pipelines.render import (
     BaseAnimatedRender,
@@ -102,10 +93,8 @@ from .pipelines.render import (
     StaticAnimatedRender,
     VideoRender,
 )
-
 # Module with classes for serializing/deserializing objects.
 from .serializer import PickleSerializer, JSONSerializer, FileJsonSerializer
-
 
 __all__ = [
     "APIMimeTyper",
@@ -145,8 +134,12 @@ __all__ = [
     "PathFromURLExtractor",
     "PickleSerializer",
     "PillowImage",
-    "Pipeline",
+    "PipelineContent",
+    "PipelineEngine",
+    "PipelineOrderedDependency",
+    "PipelineSequential",
     "Processor",
+    "ProcessorContent",
     "RarCompressedFilesFromPackageExtractor",
     "BaseRenamer",
     "BaseRender",
@@ -187,7 +180,7 @@ class ContentFile(BaseFile):
     a new one from memory using `ContentFile`.
     """
 
-    extract_data_pipeline: Pipeline = Pipeline(
+    extract_data_pipeline: PipelineEngine = PipelineOrderedDependency(
         "filejacket.pipelines.extractor.FilenameFromMetadataExtractor",
         "filejacket.pipelines.extractor.MimeTypeFromFilenameExtractor",
         "filejacket.pipelines.extractor.MimeTypeFromContentExtractor",
@@ -202,7 +195,7 @@ class StreamFile(BaseFile):
     Class to create a file from an HTTP stream that has a header with metadata.
     """
 
-    extract_data_pipeline: Pipeline = Pipeline(
+    extract_data_pipeline: PipelineEngine = PipelineOrderedDependency(
         "filejacket.pipelines.extractor.FilenameFromMetadataExtractor",
         "filejacket.pipelines.extractor.FilenameFromURLExtractor",
         "filejacket.pipelines.extractor.MimeTypeFromFilenameExtractor",
@@ -221,7 +214,7 @@ class File(BaseFile):
     a new one from memory using `ContentFile`.
     """
 
-    extract_data_pipeline: Pipeline = Pipeline(
+    extract_data_pipeline: PipelineEngine = PipelineOrderedDependency(
         "filejacket.pipelines.extractor.FilenameAndExtensionFromPathExtractor",
         "filejacket.pipelines.extractor.MimeTypeFromFilenameExtractor",
         "filejacket.pipelines.extractor.FileSystemDataExtractor",
